% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpop.R
\name{getpop}
\alias{getpop}
\title{Get local human population density}
\usage{
getpop(year = 2009, adjusted = FALSE)
}
\arguments{
\item{year}{A numeric reflecting the year to retrieve the data. Could be 2009,
2010, 2015 or 2020. 2009 is incompatible with adjusted = TRUE.}

\item{adjusted}{A logical indicating whether the data should be adjusted to
match UN population divisions estimates. TRUE is incompatible with year = 2009.}
}
\description{
Retrieves a raster file containing the local human population density of
Vietnam.
}
