% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popcenters2009}
\alias{popcenters2009}
\title{Population centers of Vietnam in 2009}
\format{An \code{sf} unprojected POINT object containing the coordinates of
the populations centers of each province together with the names of the
provinces that is \code{epi^x} compatible.}
\usage{
popcenters2009
}
\description{
An \code{sf} object containing the names and the coordinates of the
populations centers of the 63 provinces of Vietnam.
}
\details{
The provinces are as defined from the first of January 2008 and the
population centers are calculated as explained
\href{https://choisy.github.io/2018/10/29/computing-population-centers}{here},
using the 2009 population sizes.
}
\keyword{datasets}
