% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpopVN.R
\docType{package}
\name{worldpopVN}
\alias{worldpopVN}
\alias{worldpopVN-package}
\title{worldpopVN: The population densities for Vietnam from WorldPop.}
\description{
The worldpopVN package provides one rasterLayer: ppp09.
}
